+++ lib/model/auth2.dart
@@ -1819,7 +1819,7 @@ class Auth2UserPrefs {
     if (_foodFilters != null) {
       _foodFilters = null;
       if (notify == true) {
-        NotificationService().notify(notifyInterestsChanged);
+        NotificationService().notify(notifyFoodChanged);
       }
       modified = true;
     }
@@ -2253,57 +2253,84 @@ class Auth2UserPrefs {
 
   // Food
 
-  Set<String>? get excludedFoodIngredients {
-    return (_foodFilters != null) ? _foodFilters![_foodExcludedIngredients] : null;
-  }
+  Set<String>? get excludedFoodIngredients =>
+    _foodFilters?[_foodExcludedIngredients];
 
   set excludedFoodIngredients(Set<String>? value) {
-    if (!const SetEquality().equals(excludedFoodIngredients, value)) {
+    if (applyExcludedFoodIngredients(value)) {
+      NotificationService().notify(notifyFoodChanged);
+      NotificationService().notify(notifyChanged, this);
+    }
+  }
+
+  @protected
+  bool applyExcludedFoodIngredients(Set<String>? value) {
+    if (const SetEquality().equals(excludedFoodIngredients, value)) {
+      return false;
+    } else {
       if (value != null) {
         if (_foodFilters != null) {
-          _foodFilters![_foodExcludedIngredients] = value;
+          _foodFilters![_foodExcludedIngredients] = Set<String>.from(value);
         }
         else {
-          _foodFilters = { _foodExcludedIngredients : value };
+          _foodFilters = { _foodExcludedIngredients : Set<String>.from(value) };
         }
       }
       else if (_foodFilters != null) {
         _foodFilters!.remove(_foodExcludedIngredients);
       }
+      return true;
+    }
+  }
+
+  Set<String>? get includedFoodTypes =>
+    _foodFilters?[_foodIncludedTypes];
+
+  set includedFoodTypes(Set<String>? value) {
+    if (applyIncludedFoodTypes(value)) {
       NotificationService().notify(notifyFoodChanged);
       NotificationService().notify(notifyChanged, this);
     }
   }
 
-  Set<String>? get includedFoodTypes {
-    return (_foodFilters != null) ? _foodFilters![_foodIncludedTypes] : null;
-  }
-
-  set includedFoodTypes(Set<String>? value) {
-    if (!const SetEquality().equals(includedFoodTypes, value)) {
+  @protected
+  bool applyIncludedFoodTypes(Set<String>? value) {
+    if (const SetEquality().equals(includedFoodTypes, value)) {
+      return false;
+    } else {
       if (value != null) {
         if (_foodFilters != null) {
-          _foodFilters![_foodIncludedTypes] = value;
+          _foodFilters![_foodIncludedTypes] = Set<String>.from(value);
         }
         else {
-          _foodFilters = { _foodIncludedTypes : value };
+          _foodFilters = { _foodIncludedTypes : Set<String>.from(value) };
         }
       }
       else if (_foodFilters != null) {
         _foodFilters!.remove(_foodIncludedTypes);
       }
+      return true;
+    }
+  }
+
+  void applyFoodFilters({Set<String>? includedFoodTypes, Set<String>? excludedFoodIngredients}) {
+    // NB: Make sure both apply
+    bool includedFoodTypesModified = (includedFoodTypes != null) ? applyIncludedFoodTypes(includedFoodTypes) : false;
+    bool excludedFoodIngredientsModified = (excludedFoodIngredients != null) ? applyExcludedFoodIngredients(excludedFoodIngredients) : false;
+    if (includedFoodTypesModified || excludedFoodIngredientsModified) {
       NotificationService().notify(notifyFoodChanged);
       NotificationService().notify(notifyChanged, this);
     }
   }
 
-  bool get hasFoodFilters {
-    return foodFiltersNotEmpty(_foodFilters);
-  }
+  bool get hasFoodFilters =>
+    hasIncludedFoodTypes || hasExcludedIngredientsTypes;
 
-  static bool foodFiltersNotEmpty(Map<String, Set<String>>? foodFilters) {
-    return (foodFilters?[_foodIncludedTypes]?.isNotEmpty ?? false) || (foodFilters?[_foodExcludedIngredients]?.isNotEmpty ?? false);
-  }
+  bool get hasIncludedFoodTypes =>
+    _foodFilters?[_foodIncludedTypes]?.isNotEmpty == true;
+
+  bool get hasExcludedIngredientsTypes =>
+    _foodFilters?[_foodExcludedIngredients]?.isNotEmpty == true;
 
   void clearFoodFilters() {
     if (hasFoodFilters) {